refactor(gateway): extract session typing state
